Maintain built-in formatters in Builtin object (WIP)
In order to keep the interface clean for Templates there is now
a new table Builtins.

All methods in Builtins are automatically exposed as built-in commands
and can be accessed through Templates:format()'s `key` access.
They are not automatically created to be LaTeX macros but can be made so
in mapping.

They can also be access as toplevel methods for Templates (looked up in __index)

TODO: style() and shorthand() should probably be (re)moved once the macro generation
commands have been reimplemented.