Automatically check options in formatter functions
When the formatter is a function and the argument list includes
an optional argument this is intercepted and processed with
Formatter:check_options() before the function is called.
This means that within a formatter function the `options` variable
is guaranteed to be a parsed table, and if the Formatter object has an
`options` declaration the table is already prepopulated and validated.