Fix Formatter:check_option()
The handling of client options, Formatter options and local options
wasn't done correctly so far.
For each local option it has to be determined whether there is
an "option client" or if the Formatter has options defined
(these are ignored when an option client is given for the option).
In either case the corresponding Options object is used for validation,
otherwise the value is taken from the local options.